requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.6.0
coverage==6.2
Django==3.2.9
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.20.2
Faker==9.9.0
handsdown==0.4.2
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.2.1
pycodestyle==2.8.0
PyJWT==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
typed-ast==1.5.1
uritemplate==4.1.1
zipp==3.6.0
